fixup! Implement the iOS demo with shared code
