+++ kotlin-native/samples/fullstack/clients/shared/src/main/kotlin/rendering.kt
@@ -243,7 +243,7 @@ private fun loadTextureFromBmpResource(resourceName: String) {
     glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_LINEAR)
     glPixelStorei(GL_UNPACK_ALIGNMENT, 1)
 
-    val fileBytes = readResource(resourceName)
+    val fileBytes = kommon.readResource(resourceName)
 
     fileBytes.usePinned {
         val fileBytesPtr = it.addressOf(0)
