+++ kotlin-native/samples/fullstack/common/KommonIOs.kt
@@ -1,0 +1,14 @@
+package kommon
+
+import kotlinx.cinterop.*
+import objc.*
+
+fun readResource(resourceName: String): ByteArray {
+    val filePath = NSBundle.mainBundle.pathForResource(resourceName, ofType = null)
+
+    val fileData = NSData.dataWithContentsOfFile(filePath!!)
+            ?: throw Error("failed reading resource $resourceName")
+
+    return fileData.bytes!!.readBytes(fileData.length.toInt())
+}
+
